requirements.txt
# Core dependencies
streamlit
python-dotenv
groq
pydantic
concurrent-log-handler

# Utility packages
pandas
matplotlib
seaborn 
plotly

# Testing
pytest